redis integrated

+++ pkg/model/rule.go
@@ -2,5 +2,5 @@ package model
 
 type Rule struct {
 	Pattern string `yaml:"pattern"`
-	Limit   int    `yaml:"limit"`
+	Limit   int64  `yaml:"limit"`
 }
